+++ projects/igniteui-angular/src/lib/query-builder/query-builder-tree.component.ts
@@ -60,8 +60,6 @@ const DEFAULT_PIPE_DATE_FORMAT = 'mediumDate';
 const DEFAULT_PIPE_TIME_FORMAT = 'mediumTime';
 const DEFAULT_PIPE_DATE_TIME_FORMAT = 'medium';
 const DEFAULT_PIPE_DIGITS_INFO = '1.0-3';
-const DEFAULT_DATE_TIME_FORMAT = 'dd/MM/yyyy HH:mm:ss a';
-const DEFAULT_TIME_FORMAT = 'hh:mm:ss a';
 
 @Pipe({
     name: 'fieldFormatter',
@@ -820,8 +818,13 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
         const innerQuery = this.innerQueries.filter(q => q.isInEditMode())[0];
         return this.selectedField && this.selectedCondition &&
             (
-                ((!!this.searchValue.value || (!!this.searchValueTemplate && !!this._editedExpression.expression.searchVal)) && !(this.selectedField?.filters?.condition(this.selectedCondition)?.isNestedQuery)) ||
-                (innerQuery && !!innerQuery.expressionTree && innerQuery._editedExpression == undefined) ||
+                (
+                    (!!this.searchValue.value || (!!this.searchValueTemplate && !!this._editedExpression.expression.searchVal)) &&
+                    !(this.selectedField?.filters?.condition(this.selectedCondition)?.isNestedQuery)
+                ) ||
+                (
+                    innerQuery && !!innerQuery.expressionTree && innerQuery._editedExpression == undefined && innerQuery.selectedReturnFields.length > 0
+                ) ||
                 this.selectedField.filters.condition(this.selectedCondition).isUnary
             );
     }
@@ -1261,14 +1264,6 @@ export class IgxQueryBuilderTreeComponent implements AfterViewInit, OnDestroy {
             field.pipeArgs.format = field.dataType === DataType.Time ?
                 DEFAULT_PIPE_TIME_FORMAT : field.dataType === DataType.DateTime ?
                     DEFAULT_PIPE_DATE_TIME_FORMAT : DEFAULT_PIPE_DATE_FORMAT;
-        }
-
-        if (!field.defaultDateTimeFormat) {
-            field.defaultDateTimeFormat = DEFAULT_DATE_TIME_FORMAT;
-        }
-
-        if (!field.defaultTimeFormat) {
-            field.defaultTimeFormat = DEFAULT_TIME_FORMAT;
         }
     }